requirements.txt
asgiref==3.4.1
astroid==2.8.0
Cerberus==1.3.4
certifi==2021.5.30
charset-normalizer==2.0.6
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.3.0
psycopg2==2.9.1
pylint==2.11.1
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
